[14.0] mail_debrand: Split code into model-specific files
