+++ mail_debrand/models/__init__.py
@@ -1,1 +1,2 @@
-from . import mail_render_mixinANDmail_mail
+from . import mail_render_mixin
+from . import mail_mail
